Move connection settings to human class
